chore(postgres): share db creation pool code
