+++ pkg/plugins/common/postgres/pgx_listener.go
@@ -34,11 +34,7 @@ var _ Listener = (*PgxListener)(nil)
 // NewPgxListener will create and initialize a PgxListener which will automatically connect and listen to the provided channel.
 func NewPgxListener(config postgres.PostgresStoreConfig, logger logr.Logger) (Listener, error) {
 	ctx := context.Background()
-	connectionString, err := config.ConnectionString()
-	if err != nil {
-		return nil, err
-	}
-	db, err := pgxpool.New(context.Background(), connectionString)
+	db, err := ConnectToDbPgx(config)
 	if err != nil {
 		return nil, err
 	}
